producer/requirements.txt
kafka-python==2.0.2
clickhouse-connect==0.8.15
